@@ -1,13 +1,16 @@
 package gr.teicm.ieee.madc.disasternotifierandroid;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+
+//import android.support.test.InstrumentationRegistry;
 
 /**
  * Instrumented test, which will execute on an Android device.
